+++ ng-dev/pr/merge/config.ts
@@ -19,23 +19,7 @@ export type TargetLabelBranchResult = string[] | Promise<string[]>;
  */
 export type GithubApiMergeMethod = 'merge' | 'squash' | 'rebase';
 
-/**
- * Target labels represent Github pull requests labels. These labels instruct the merge
- * script into which branches a given pull request should be merged to.
- */
-export interface TargetLabel {
-  /** Pattern that matches the given target label. */
-  pattern: RegExp | string;
-  /**
-   * List of branches a pull request with this target label should be merged into.
-   * Can also be wrapped in a function that accepts the target branch specified in the
-   * Github Web UI. This is useful for supporting labels like `target: development-branch`.
-   *
-   * @throws {InvalidTargetLabelError} Invalid label has been applied to pull request.
-   * @throws {InvalidTargetBranchError} Invalid Github target branch has been selected.
-   */
-  branches: TargetLabelBranchResult | ((githubTargetBranch: string) => TargetLabelBranchResult);
-}
+
 
 /**
  * Configuration for the merge script with all remote options specified. The
